+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 -   Re-enabled node reward program. Upgraded TLS communication.
 -   Fixed `failed to load: /docker-entrypoint-initdb.d/mongoDbPrepare.js` when running with root user.
 -   Renamed `enrol` to `enroll` for the Rewards Program.
+-   Added `wizard` command.
+-   Added `pack` command.
 
 ## [1.0.4] - Apr-13-2021
 
@@ -24,8 +26,6 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 | ---------------- | ------- | ------------------------------------------------------------------ |
 | Symbol Bootstrap | v1.0.4  | [symbol-bootstrap](https://www.npmjs.com/package/symbol-bootstrap) |
 
--   Added `wizard` command.
--   Added `pack` command.
 -   New `testnet.symboldev.network` testnet!
 -   Added `verify` command.
 -   Fixed host override when no custom preset is provided in mainnet.